feat: verify use of ingredients before delete

+++ src/main/java/de/slevermann/cocktails/backend/service/IngredientService.java
@@ -5,8 +5,9 @@ import de.slevermann.cocktails.api.model.Ingredient;
 import de.slevermann.cocktails.backend.dao.IngredientDao;
 import de.slevermann.cocktails.backend.dao.IngredientTypeDao;
 import de.slevermann.cocktails.backend.model.mapper.IngredientMapper;
-import de.slevermann.cocktails.backend.service.problem.BadReferenceProblem;
+import de.slevermann.cocktails.backend.service.problem.MissingReferenceProblem;
 import de.slevermann.cocktails.backend.service.problem.NoSuchResourceProblem;
+import de.slevermann.cocktails.backend.service.problem.ReferencedEntityProblem;
 import de.slevermann.cocktails.backend.service.problem.ResourceType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -40,13 +41,22 @@ public class IngredientService {
         final var typeId = createIngredient.getType();
         final var type = ingredientTypeDao.getById(typeId);
         if (type == null) {
-            throw new BadReferenceProblem(ResourceType.INGREDIENT_TYPE, typeId.toString());
+            throw new MissingReferenceProblem(ResourceType.INGREDIENT_TYPE, typeId.toString());
         }
         return ingredientMapper.fromDb(
                 ingredientDao.create(ingredientMapper.fromApi(createIngredient)));
     }
 
+    @Transactional
     public void delete(final UUID uuid) {
+        if (ingredientDao.usedByCount(uuid) > 0) {
+            throw new ReferencedEntityProblem(ResourceType.INGREDIENT, uuid.toString());
+        }
+
+        if (ingredientDao.shelfCount(uuid) > 0) {
+            throw new ReferencedEntityProblem(ResourceType.INGREDIENT, uuid.toString());
+        }
+
         if (ingredientDao.delete(uuid) == 0) {
             throw new NoSuchResourceProblem(ResourceType.INGREDIENT, uuid.toString());
         }
